Connect ImportComponents to editor slice
