+++ ui/src/components/editor/PlateSidebar/ComponentList/ImportComponents/ImportComponents.js
@@ -1,9 +1,7 @@
 import React, { Component } from 'react';
 import PropTypes from 'prop-types';
-import { connect } from 'react-redux';
 
 import { api } from '../../../../../api';
-import { addKaptureComponentsToComponents } from '../../../../../store/designActions';
 import { ImportTextArea } from './ImportTextArea';
 import { ImportResults } from './ImportResults';
 
@@ -17,7 +15,7 @@ const INITIAL_STATE = {
   completed: false,
 };
 
-class ImportComponents extends Component {
+export class ImportComponents extends Component {
   componentNames = [];
   currentSearchIndex = 0;
   state = INITIAL_STATE;
@@ -92,13 +90,3 @@ ImportComponents.propTypes = {
   onAdd: PropTypes.func,
   afterAdd: PropTypes.func,
 };
-
-const mapDispatch = {
-  onAdd: addKaptureComponentsToComponents,
-};
-
-const connected = connect(
-  null,
-  mapDispatch
-)(ImportComponents);
-export { connected as ImportComponents };
